# swagger/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "v1",
    "title" : "API"
  },
  "host" : "www.mannetroll.com",
  "basePath" : "/eswebbapp",
...
  "schemes" : [ "https" ],
  "paths" : {
    "/arrivalcompleted/spid/{spid}" : {
      "get" : {
        "tags" : [ "ArrivalCompleted" ],
        "summary" : "arrivalcompleted",
        "description" : "notes",
        "operationId" : "arrivalCompleted",
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "name" : "spid",
          "in" : "path",
          "description" : "ServicePointID",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "percentile",
          "in" : "query",
          "description" : "Percentile",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ArrivalCompletedResponse"
            }
          }
        }
      }
    },
    "/arrivalcompleted/spid/{spid}/dayofweek/{dayofweek}" : {
      "get" : {
        "tags" : [ "ArrivalCompleted" ],
        "summary" : "arrivalcompleteddayofweek",
        "description" : "notes",
        "operationId" : "arrivalCompletedDayOfWeek",
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "name" : "spid",
          "in" : "path",
          "description" : "ServicePointID",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "dayofweek",
          "in" : "path",
          "description" : "DayOfWeek",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "percentile",
          "in" : "query",
          "description" : "Percentile",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ArrivalCompletedResponse"
            }
          }
        }
      }
    },
    "/arrivalcompleted/spid/{spid}/percentile/{percentile}/date/{date}/weeks/{weeks}" : {
      "get" : {
        "tags" : [ "ArrivalCompleted" ],
        "summary" : "arrivalcompletedparam",
        "description" : "notes",
        "operationId" : "arrivalCompletedParam",
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "name" : "spid",
          "in" : "path",
          "description" : "ServicePointID",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "percentile",
          "in" : "path",
          "description" : "percentile",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "date",
          "in" : "path",
          "description" : "date",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "weeks",
          "in" : "path",
          "description" : "weeks",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ArrivalCompletedResponse"
            }
          }
        }
      }
    },
    "/arrivalcompleted2/spid/{spid}" : {
      "get" : {
        "tags" : [ "ArrivalCompleted" ],
        "summary" : "arrivalcompleted2",
        "description" : "notes",
        "operationId" : "arrivalCompleted2",
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "name" : "spid",
          "in" : "path",
          "description" : "ServicePointID",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ArrivalCompletedResponse"
            }
          }
        }
      }
    },
    "/arrivalcompleted2/spid/{spid}/dayofweek/{dayofweek}" : {
      "get" : {
        "tags" : [ "ArrivalCompleted" ],
        "summary" : "arrivalcompleteddayofweek2",
        "description" : "notes",
        "operationId" : "arrivalCompletedDayOfWeek2",
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "name" : "spid",
          "in" : "path",
          "description" : "ServicePointID",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "dayofweek",
          "in" : "path",
          "description" : "DayOfWeek",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ArrivalCompletedResponse"
            }
          }
        }
      }
    },
...
    "/arrivalcompletedexport" : {
      "get" : {
        "tags" : [ "ArrivalCompleted" ],
        "summary" : "arrivalcompletedexport",
        "description" : "notes",
        "operationId" : "arrivalCompletedExport",
        "produces" : [ "text/plain" ],
        "parameters" : [ {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ArrivalCompletedResponse"
            }
          }
        }
      }
    },
    "/arrivalcompletedlist/{spid}" : {
      "get" : {
        "tags" : [ "ArrivalCompleted" ],
        "summary" : "arrivalcompletedlist",
        "description" : "notes",
        "operationId" : "arrivalCompletedIds",
        "produces" : [ "text/plain" ],
        "parameters" : [ {
          "name" : "spid",
          "in" : "path",
          "description" : "ServicePointID",
          "required" : true,
          "type" : "array",
          "items" : {
            "type" : "string"
          },
          "collectionFormat" : "csv"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ArrivalCompletedResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
    "ApiError" : {
      "type" : "object",
      "required" : [ "faults" ],
      "properties" : {
        "faults" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Fault"
          }
        }
      },
      "description" : "Model for API errors"
    },
    "ArrivalCompletedResponse" : {
      "type" : "object",
      "properties" : {
        "error" : {
          "type" : "string"
        },
        "apiError" : {
          "$ref" : "#/definitions/ApiError"
        },
        "name" : {
          "type" : "string"
        },
        "servicePointId" : {
          "type" : "string"
        },
        "arrivalCompleted" : {
          "type" : "array",
          "items" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object"
            }
          }
        },
        "percentile" : {
          "type" : "integer",
          "format" : "int32"
        },
        "weeks" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "Fault" : {
      "type" : "object",
      "required" : [ "explanationText", "faultCode" ],
      "properties" : {
        "faultCode" : {
          "type" : "string"
        },
        "explanationText" : {
          "type" : "string"
        },
        "paramValues" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ParamValues"
          }
        }
      },
      "description" : "Model for API Fault"
    },
...
    "ParamValues" : {
      "type" : "object",
      "required" : [ "param", "value" ],
      "properties" : {
        "param" : {
          "type" : "string"
        },
        "value" : {
          "type" : "string"
        }
      },
      "description" : "Key/Value information about errors"
    },
...
  }
}